+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Added
 
+- Add Tooltip element
 - Support arrow in Floating panel element
 - Allow trigger id to be set in the Floating panel element